Make generator "resumable"
When the generator is (re)started, it looks for Deployments in the
requested namespace that have the ocs-monkey/controller=osio label,
meaning they were started by this generator. For each one it finds, it
schedules a Lifecycle event to resume managing it.

This functionality was added in preparation for running the generator in
the OCP cluster, where it must withstand disruptions/restarts.

Signed-off-by: John Strunk <jstrunk@redhat.com>