WORKAROUND: Ensure workload pods have min runtime
This workaround ensures each workload instance spends some minimum
amount of time in each active/idle state before transitioning. Currently
this is fixed at the initial_health_interval. This workaround is
necessary due to a CSI bug that leaves volumes attached to nodes if a
request is canceled before the attach succeeds.

Signed-off-by: John Strunk <jstrunk@redhat.com>